Refactor form modules
